Add instance var for validating modifiers
Adds new instance variable `validateModifiers` for disabling modifiers validation.
It also adds a new initializer to the public API to create a new view with the setting
set to `true` or `false`. Setting `validateModifiers` to `false` will disable all
types of validation of modifiers and thus adds support for recording single letter
shortcuts without modifiers. The default value is `true`.